Move companies to a new folder
